+++ Source/AccelByteUe4Sdk/Public/Models/AccelByteOauth2Models.h
@@ -111,6 +111,12 @@ struct ACCELBYTEUE4SDK_API FOauth2Token
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "AccelByte | Oauth2 | Models | Oauth2Token")
 	FString Unique_display_name{};
+
+	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "AccelByte | Oauth2 | Models | Oauth2Token")
+	FString Simultaneous_platform_id{};
+	
+	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "AccelByte | Oauth2 | Models | Oauth2Token")
+	FString Simultaneous_platform_user_id{};
 };
 
 USTRUCT(BlueprintType)